requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
sentence-transformers
llama-cpp-python